GDF-129_2 Travis + Vagrant

+++ puppet/initial/init.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
-dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
-ip="$(ip addr | grep 'state UP' -A2 | tail -n1 | awk '{print $2}' | cut -f1  -d'/')";
-
-role="local"
+modules_dir=$(dirname $0)/../modules
+cd ${modules_dir}/../..
+project_dir=$(pwd)
+daemon_user='www-data'
 
 show_help()
 {
@@ -10,21 +10,19 @@ cat << EOF
 This script download files from remote server and download tar.gz to local machine or remote host.
 Usage: $0 options
 OPTIONS:
-    -t  github token
+    -u  daemon user
     -h  show this message
 EOF
 }
 
-
-while getopts "t:h:r:" OPTION
+while getopts "u:h:" OPTION
 do
      case ${OPTION} in
+         u)
+             daemon_user=$OPTARG
+             ;;
          h)
              show_help
-             exit 1
-             ;;
-         r)
-             role=$OPTARG
              ;;
          ?)
              show_help
@@ -33,26 +31,42 @@ do
      esac
 done
 
+if [ ! -e /usr/bin/puppet ]; then
+    source /etc/lsb-release
+    wget https://apt.puppetlabs.com/puppetlabs-release-$DISTRIB_CODENAME.deb
+    sudo dpkg -i puppetlabs-release-$DISTRIB_CODENAME.deb
+    rm puppetlabs-release-$DISTRIB_CODENAME.deb
+    sudo apt-get update
+    sudo apt-get install -y -f puppet git
+    if [ ! -d "/etc/puppet/environments" ]; then
+        sudo mkdir /etc/puppet/environments;
+    fi
+    sudo chgrp puppet /etc/puppet/environments
+    sudo chmod 2775 /etc/puppet/environments
+    echo '
+    START=yes
+    DAEMON_OPTS=""
+    ' | sudo tee --append /etc/default/puppet
+    sudo service puppet start
+else
+    #echo 1;
+    sudo apt-get update
+fi;
 
-source /etc/lsb-release
-wget https://apt.puppetlabs.com/puppetlabs-release-$DISTRIB_CODENAME.deb
-sudo dpkg -i puppetlabs-release-$DISTRIB_CODENAME.deb
-rm puppetlabs-release-$DISTRIB_CODENAME.deb
-sudo apt-get update
-sudo apt-get install -y -f puppet git
-if [ ! -d "/etc/puppet/environments" ]; then
-    sudo mkdir /etc/puppet/environments;
-fi
-sudo chgrp puppet /etc/puppet/environments
-sudo chmod 2775 /etc/puppet/environments
-echo '
-START=yes
-DAEMON_OPTS=""
-' | sudo tee --append /etc/default/puppet
-sudo service puppet start
-sudo mkdir -p /www/
-sudo chown www-data:www-data /www/
-sudo mkdir -p /var/www/.ssh
-sudo chown -Rf www-data:www-data /var/www/
+if [ ! -e /www ]; then
+    sudo mkdir /www/
+    sudo chmod 755 /www/
+    sudo chown www-data:www-data /www/
+    sudo mkdir -p /var/www/.ssh
+    sudo chown -Rf www-data:www-data /var/www/
+fi;
 
-sudo FACTER_server_tags="role:${role}" puppet apply --modulepath /www/gandalf.api/puppet/modules /www/gandalf.api/puppet/general/manifests/init.pp
+sudo puppet module install --force puppetlabs/stdlib --target-dir ${modules_dir}
+sudo puppet module install --force puppetlabs/apt --target-dir ${modules_dir}
+sudo puppet module install --force maestrodev/wget --target-dir ${modules_dir}
+sudo puppet module install --force willdurand/composer --target-dir ${modules_dir}
+sudo puppet module install --force jfryman-nginx --target-dir ${modules_dir}
+sudo puppet module install --force saz-timezone --target-dir ${modules_dir}
+sudo puppet module install --force saz-locales --target-dir ${modules_dir}
+
+sudo FACTER_project_dir="${project_dir}" FACTER_daemon_user="${daemon_user}" puppet apply --modulepath ${modules_dir} ${modules_dir}/../general/manifests/init.pp
